requirements.txt
# Dependências principais
flask==2.3.3
flask-cors==4.0.0
flask-caching==2.1.0
python-dotenv==1.0.0
requests==2.31.0
pytz==2023.3

# OpenAI SDK
openai==1.3.7

# Supabase
supabase==2.0.2

# Redis (para cache)
redis==5.0.1

# Celery para tarefas assíncronas
celery==5.3.4

# Flower para monitoramento do Celery
flower==2.0.1

# SFTP e XML parsing
paramiko==3.4.0
xmltodict==0.13.0

# Utilitários de data
python-dateutil==2.8.2

# Para geração de PDF
reportlab==4.0.4
